Move o.e.painless.toxcontent.UserTreeToXContent to test codebase
This thing is only used in tests we can move it there.
